checking if dir exists before deleting

+++ scripts/clean-local.sh
@@ -12,21 +12,24 @@ echo "rm -rf ~/workspace/pn/visitlocal/"
 rm -rf ~/workspace/pn/visitlocal/
 
 #cleaning hardhat cache.
-echo "cleaning hardhat cache and temp files"
-cd ../point-contracts
-echo "rm -rf cache"
-rm -rf cache
-echo "rm -rf typechain"
-rm -rf typechain
-echo "rm -rf build"
-rm -rf build
-echo "rm -rf .openzeppelin"
-rm -rf .openzeppelin
-echo "rm resources/Identity-address.json"
-rm -f resources/Identity-address.json
-echo "rm resources/unknown-1337.json"
-rm -f resources/unknown-1337.json
-cd ../pointnetwork
+echo "cleaning point-contracts cache and temp files"
+
+if [ -d ../point-contracts ]; then
+    cd ../point-contracts
+    echo "rm -rf cache"
+    rm -rf cache
+    echo "rm -rf typechain"
+    rm -rf typechain
+    echo "rm -rf build"
+    rm -rf build
+    echo "rm -rf .openzeppelin"
+    rm -rf .openzeppelin
+    echo "rm resources/Identity-address.json"
+    rm -f resources/Identity-address.json
+    echo "rm resources/unknown-1337.json"
+    rm -f resources/unknown-1337.json
+    cd ../pointnetwork
+fi
 
 #removing log files
 echo "rm *.log"
